@@ -6,8 +6,13 @@ namespace Ministry.Web.Themes
     /// <summary>
     /// Options that link theme keys and hosts together.
     /// </summary>
-    [SuppressMessage("ReSharper", "ClassNeverInstantiated.Global")]
-    [SuppressMessage("ReSharper", "UnusedMember.Global")]
+    /// <seealso cref="IKeyedTheme"/>
+    /// <inheritdoc cref="IKeyedTheme"/>
+    [SuppressMessage("ReSharper", "ClassNeverInstantiated.Global", Justification = "Library")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global", Justification = "Library")]
+    [SuppressMessage("ReSharper", "CollectionNeverUpdated.Global", Justification = "Library")]
+    [SuppressMessage("ReSharper", "AutoPropertyCanBeMadeGetOnly.Global", Justification = "Library")]
+    [SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global", Justification = "Library")]
     public class HostTheme : IKeyedTheme
     {
         #region | Construction |
@@ -22,15 +27,13 @@ namespace Ministry.Web.Themes
 
         #endregion
 
+        /// <inheritdoc/>>
+        public string Key { get; set; }
+
         /// <summary>
         /// Gets or Sets the Hosts that are linked to this theme.
         /// </summary>
         public List<string> ValidHosts { get; set; }
-
-        /// <summary>
-        /// Gets or sets the theme key.
-        /// </summary>
-        public string Key { get; set; }
 
         /// <summary>
         /// Gets or sets the theme title.
