@@ -1,8 +1,11 @@
-﻿namespace Ministry.Web.Themes
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Ministry.Web.Themes
 {
     /// <summary>
     /// Representation of a theme with a key.
     /// </summary>
+    [SuppressMessage("ReSharper", "UnusedMemberInSuper.Global", Justification = "Library")]
     public interface IKeyedTheme
     {
         /// <summary>
